# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AudiModal.ai API",
    "description": "Comprehensive AI-powered document processing and analytics platform with ML/AI classification, anomaly detection, unified sync framework, and enterprise integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "AudiModal.ai",
      "url": "https://audimodal.ai",
      "email": "support@audimodal.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.audimodal.ai/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.audimodal.ai/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/tenants": {
      "get": {
        "tags": ["Tenants"],
        "summary": "List tenants",
        "description": "Retrieve a list of tenants with optional filtering and pagination",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of tenants to return",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of tenants to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tenants": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tenant"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tenants"],
        "summary": "Create tenant",
        "description": "Create a new tenant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}": {
      "get": {
        "tags": ["Tenants"],
        "summary": "Get tenant",
        "description": "Retrieve a specific tenant by ID",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          }
        }
      },
      "put": {
        "tags": ["Tenants"],
        "summary": "Update tenant",
        "description": "Update an existing tenant",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tenants"],
        "summary": "Delete tenant",
        "description": "Delete a tenant and all associated data",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tenant deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Tenant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "suspended", "inactive"]
          },
          "settings": {
            "type": "object",
            "properties": {
              "max_storage_gb": {
                "type": "integer"
              },
              "max_users": {
                "type": "integer"
              },
              "retention_days": {
                "type": "integer"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTenantRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "settings": {
            "type": "object",
            "properties": {
              "max_storage_gb": {
                "type": "integer",
                "default": 100
              },
              "max_users": {
                "type": "integer",
                "default": 50
              },
              "retention_days": {
                "type": "integer",
                "default": 365
              }
            }
          }
        },
        "required": ["name", "domain"]
      },
      "UpdateTenantRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "suspended", "inactive"]
          },
          "settings": {
            "type": "object",
            "properties": {
              "max_storage_gb": {
                "type": "integer"
              },
              "max_users": {
                "type": "integer"
              },
              "retention_days": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    }
  },
...
}